+++ lib/lua/src/lua_box.cpp
@@ -1,0 +1,7 @@
+#include "lua_box.hpp"
+
+LuaBox::LuaBox(LuaWidget* parent, int x, int y, int width, int height)
+{
+    widget = new Box(x, y, width, height);
+    init(widget, parent, x, y, width, height);
+}
