+++ lib/lua/src/lua_file.cpp
@@ -7,467 +7,6 @@
 #include <threads.hpp>
 
 
-void LuaWidget::init(gui::ElementBase* obj, LuaWidget* parent, int x, int y, int width, int height)
-{
-    widget = obj;
-    parent->addChild(this);
-    parent->widget->addChild(this->widget);
-}
-
-LuaWidget::~LuaWidget()
-{
-    for (LuaWidget* child : children)
-    {
-        delete child;
-    }
-
-    if (this->widget->getParent() == nullptr)
-    {
-        delete this->widget;
-        this->widget = nullptr;
-    }
-}
-
-void LuaWidget::update()
-{
-    if(this->widget->getParent() == nullptr)
-        this->widget->updateAll();
-
-    if(onClickFunc && this->widget->isTouched())
-    {
-        std::cout << "onClickFunc" << std::endl;
-        onClickFunc();
-    }
-
-    /*if(onScrollUpFunc && touch.isSlidingVertically() < -30)
-    {
-        onScrollUpFunc();
-        touch.resetScrollVertical();
-    }
-
-    if(onScrollDownFunc && touch.isSlidingVertically() > 30)
-    {
-        onScrollDownFunc();
-        touch.resetScrollVertical();
-    }
-
-    if(onScrollRightFunc && touch.isSlidingHorizontally() > 30)
-    {
-        onScrollRightFunc();
-        touch.resetScrollHorizontal();
-    }
-
-    if(onScrollLeftFunc && touch.isSlidingHorizontally() < -30)
-    {
-        onScrollLeftFunc();
-        touch.resetScrollHorizontal();
-    }*/
-
-    for (int i = 0; i < this->children.size(); i++)
-    {
-        children[i]->update();
-    }
-}
-
-// Box
-
-LuaBox::LuaBox(LuaWidget* parent, int x, int y, int width, int height)
-{
-    widget = new Box(x, y, width, height);
-    init(widget, parent, x, y, width, height);
-}
-
-// Image
-
-LuaImage::LuaImage(LuaWidget* parent, storage::Path path, int x, int y, int width, int height)
-{
-    widget = new Image(path, x, y, width, height);
-    init(widget, parent, x, y, width, height);
-    widget->load();
-}
-
-// Label
-
-LuaLabel::LuaLabel(LuaWidget* parent, int x, int y, int width, int height)
-{
-    widget = new Label(x, y, width, height);
-    init(widget, parent, x, y, width, height);
-}
-
-// Button
-
-LuaButton::LuaButton(LuaWidget* parent, int x, int y, int width, int height)
-{
-    widget = new Button(x, y, width, height);
-    init(widget, parent, x, y, width, height);
-}
-
-// Input
-
-LuaInput::LuaInput(LuaWidget* parent, int x, int y, int width, int height)
-{
-    widget = new Input(x, y, width, height);    // a mettre a jour sur le paxos9
-    init(widget, parent, x, y, width, height);
-}
-
-
-// Window
-
-LuaWindow::LuaWindow()  
-{
-    widget = new gui::elements::Window();
-    init(widget);
-}
-
-// gui::ElementBase
-
-LuaGui::LuaGui(LuaFile* lua)
-{
-    this->lua = lua;
-}
-
-LuaBox* LuaGui::box(LuaWidget* parent, int x, int y, int width, int height)
-{
-    LuaBox* w = new LuaBox(parent, x, y, width, height);
-    return w;
-}
-
-LuaImage* LuaGui::image(LuaWidget* parent, storage::Path path, int x, int y, int width, int height)
-{
-    // TODO: permissions relatives a l'app
-    if(!this->lua->perms.acces_files)
-        return nullptr;
-    if(path.m_steps[0]=="/" && !this->lua->perms.acces_files_root)
-        return nullptr;
-    
-    LuaImage* w = new LuaImage(parent, path, x, y, width, height);
-
-    return w;
-}
-
-LuaLabel* LuaGui::label(LuaWidget* parent, int x, int y, int width, int height)
-{
-    LuaLabel* w = new LuaLabel(parent, x, y, width, height);
-    return w;
-}
-
-LuaInput* LuaGui::input(LuaWidget* parent, int x, int y, int width, int height)
-{
-    LuaInput* w = new LuaInput(parent, x, y, width, height);
-    return w;
-}
-
-LuaButton* LuaGui::button(LuaWidget* parent, int x, int y, int width, int height)
-{
-    LuaButton* w = new LuaButton(parent, x, y, width, height);
-    return w;
-}
-
-LuaWindow* LuaGui::window()
-{
-    return new LuaWindow();
-}
-
-void LuaGui::update()
-{
-    if(mainWindow != nullptr)
-    {
-        mainWindow->update();
-    }
-}
-
-// gui::ElementBase end
-
-void LuaHardware::flash(bool flash)
-{
-    std::cout << "flash" << flash << std::endl;
-}
-
-// Storage
-
-LuaStorageFile::LuaStorageFile(storage::Path filename, bool mode)
-{
-    std::cout << filename.str() << std::endl;
-    this->filename = filename;
-    this->mode = mode;
-}
-
-void LuaStorageFile::open()
-{
-    file.open(filename.str(), (mode)?(storage::WRITE):(storage::READ));
-}
-
-void LuaStorageFile::close()
-{
-    file.close();
-}
-
-void LuaStorageFile::write(const std::string& text)
-{
-    file.write(text);
-}
-
-char LuaStorageFile::readChar()
-{
-    return file.readchar();
-}
-
-std::string LuaStorageFile::readLine()
-{
-    return file.readline();
-}
-
-std::string LuaStorageFile::readAll()
-{
-    return file.read();
-}
-
-LuaStorage::LuaStorage(LuaFile* lua)
-{
-    this->lua = lua;
-}
-
-LuaStorageFile* LuaStorage::file(storage::Path filename, bool mode)
-{
-    if (!legalPath(filename))
-        return nullptr;
-
-    if(filename.m_steps[0]=="/")
-        return new LuaStorageFile(filename, mode);
-    else
-        return new LuaStorageFile(this->lua->directory / filename, mode);
-}
-
-bool LuaStorage::newDir(storage::Path path)
-{
-    return convertPath(path).newdir();
-}
-
-bool LuaStorage::renameFile(storage::Path oldpath, storage::Path newpath)
-{
-    return convertPath(oldpath).rename(convertPath(newpath));
-}
-
-bool LuaStorage::renameDir(storage::Path oldpath, storage::Path newpath)
-{
-    return convertPath(oldpath).rename(convertPath(newpath));
-}
-
-bool LuaStorage::deleteFile(storage::Path path)
-{
-    return convertPath(path).remove();
-}
-
-bool LuaStorage::deleteDir(storage::Path path)
-{
-    return convertPath(path).remove();
-}
-
-bool LuaStorage::isFile(storage::Path path)
-{
-    return convertPath(path).isfile();
-}
-
-bool LuaStorage::isDir(storage::Path path)
-{
-    return convertPath(path).isdir();
-}
-
-
-std::vector<std::string> LuaStorage::listDir(storage::Path path)
-{
-    return convertPath(path).listdir();
-}
-
-bool LuaStorage::legalPath(storage::Path path)
-{
-    if(!this->lua->perms.acces_files)
-        return false;
-    if(path.m_steps[0]=="/" && !this->lua->perms.acces_files_root)
-        return false;
-    
-    return true;
-}
-
-storage::Path LuaStorage::convertPath(storage::Path path)
-{
-    if (!legalPath(path))
-        return this->lua->directory;
-    
-    if(path.m_steps[0]=="/")
-        return path;
-    else
-        return this->lua->directory / path;
-}
-
-// end storage
-
-LuaTime::LuaTime(LuaFile* lua)
-{
-    this->lua = lua;
-    timerFromStart = millis();
-}
-
-uint32_t LuaTime::monotonic()
-{
-    return millis() - timerFromStart;
-}
-
-int findIndex(const std::vector<std::string>& vec, const std::string& target) {
-    for (std::size_t i = 0; i < vec.size(); ++i) {
-        if (vec[i] == target) {
-            return static_cast<int>(i); // Convertir en int si nécessaire
-        }
-    }
-    return -1; // Retourner -1 si la chaîne n'est pas trouvée
-}
-
-sol::table LuaTime::get(std::string format)
-{
-    // découper le format en identifiers
-    const std::string delimiter = ",";
-    std::vector<std::string> result;
-    std::string::size_type start = 0;
-    std::string::size_type end = format.find(delimiter);
-
-    while (end != std::string::npos) {
-        result.push_back(format.substr(start, end - start));
-        start = end + 1;
-        end = format.find(delimiter, start);
-    }
-
-    result.push_back(format.substr(start));
-
-    std::vector<std::string> identifiers = {"s","mi","h","d","mo","y"};
-    //std::vector<int> date = {gsm.seconds,gsm.minutes,gsm.hours,gsm.days,gsm.months,gsm.years};
-    std::vector<int> date = {0,0,0,0,0,0};
-
-    
-    // ajouter les valeurs aux index des identifiers
-
-    sol::table array = lua->lua.create_table();
-
-    // Remplir le tableau avec des nombres
-    for (int i = 0; i < result.size(); i++) {
-        int index = findIndex(identifiers, result[i]);
-        if(index!=-1)
-            array[i+1] = date[index];
-        else
-            return lua->lua.create_table();
-    }
-
-    // Retourner le tableau Lua
-    return array;
-}
-
-LuaTimeInterval::LuaTimeInterval(LuaFile* lua, sol::function func, uint32_t interval)
-{
-    this->lua = lua;
-    this->func = func;
-    this->interval = interval;
-    this->id = eventHandlerApp.setInterval(new Callback<>(std::function<void(void)>(std::bind(&LuaTimeInterval::call, this))), interval);
-}
-
-int LuaTimeInterval::getId()
-{
-    return id;
-}
-
-void LuaTimeInterval::call()
-{
-    if(func)
-        func();
-}
-
-LuaTimeInterval::~LuaTimeInterval()
-{
-    eventHandlerApp.removeInterval(id);
-}
-
-LuaTimeTimeout::LuaTimeTimeout(LuaFile* lua, sol::function func, uint32_t timeout)
-{
-    this->lua = lua;
-    this->func = func;
-    this->timeout = timeout;
-    this->id = eventHandlerApp.setTimeout(new Callback<>(std::function<void(void)>(std::bind(&LuaTimeTimeout::call, this))), timeout);
-}
-
-int LuaTimeTimeout::getId()
-{
-    return id;
-}
-
-void LuaTimeTimeout::call()
-{
-    if(func)
-        func();
-    done = true;
-}
-
-LuaTimeTimeout::~LuaTimeTimeout()
-{
-    if(!done)
-        eventHandlerApp.removeTimeout(id);
-}
-
-void LuaTime::update()
-{
-    for (int it = 0; it < timeouts.size(); it++)
-    {
-        if (timeouts[it]->done)
-        {
-            delete timeouts[it];
-            timeouts.erase(timeouts.begin() + it);
-            it = 0; // reset the loop
-        }
-    }
-}
-
-int LuaTime::setInterval(sol::function func, uint32_t interval)
-{
-    LuaTimeInterval* n = new LuaTimeInterval(lua, func, interval);
-    intervals.push_back(n);
-    return n->getId();
-}
-
-int LuaTime::setTimeout(sol::function func, uint32_t interval)
-{
-    LuaTimeTimeout* n = new LuaTimeTimeout(lua, func, interval);
-    timeouts.push_back(n);
-    return n->getId();
-}
-
-void LuaTime::removeInterval(int id)
-{
-    for (int it = 0; it < intervals.size(); it++)
-    {
-        if (intervals[it]->getId() == id)
-        {
-            delete intervals[it];
-            intervals.erase(intervals.begin() + it);
-        }
-    }
-}
-
-void LuaTime::removeTimeout(int id)
-{
-    for (int it = 0; it < timeouts.size(); it++)
-    {
-        if (timeouts[it]->getId() == id)
-        {
-            delete timeouts[it];
-            timeouts.erase(timeouts.begin() + it);
-        }
-    }
-}
-
-LuaTime::~LuaTime()
-{
-    for (int i = 0; i < intervals.size(); i++)
-        delete intervals[i];
-    for (int i = 0; i < intervals.size(); i++)
-        delete timeouts[i];
-}
 /*
 LuaHttpClient::LuaHttpClient(LuaFile* lua)
     : httpClient()
