+++ lib/lua/src/lua_hardware.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+
+#include "lua_hardware.hpp"
+
+void LuaHardware::flash(bool flash)
+{
+    std::cout << "flash" << flash << std::endl;
+}
