+++ lib/lua/src/lua_hardware.hpp
@@ -1,0 +1,10 @@
+#ifndef LUA_MODULE_HARDWARE
+#define LUA_MODULE_HARDWARE
+
+class LuaHardware
+{
+    public:
+    void flash(bool flash);
+};
+
+#endif
