+++ lib/lua/src/lua_widget.cpp
@@ -1,0 +1,63 @@
+#include "lua_widget.hpp"
+
+void LuaWidget::init(gui::ElementBase* obj, LuaWidget* parent, int x, int y, int width, int height)
+{
+    widget = obj;
+    parent->addChild(this);
+    parent->widget->addChild(this->widget);
+}
+
+LuaWidget::~LuaWidget()
+{
+    for (LuaWidget* child : children)
+    {
+        delete child;
+    }
+
+    if (this->widget->getParent() == nullptr)
+    {
+        delete this->widget;
+        this->widget = nullptr;
+    }
+}
+
+void LuaWidget::update()
+{
+    if(this->widget->getParent() == nullptr)
+        this->widget->updateAll();
+
+    if(onClickFunc && this->widget->isTouched())
+    {
+        std::cout << "onClickFunc" << std::endl;
+        onClickFunc();
+    }
+
+    /*if(onScrollUpFunc && touch.isSlidingVertically() < -30)
+    {
+        onScrollUpFunc();
+        touch.resetScrollVertical();
+    }
+
+    if(onScrollDownFunc && touch.isSlidingVertically() > 30)
+    {
+        onScrollDownFunc();
+        touch.resetScrollVertical();
+    }
+
+    if(onScrollRightFunc && touch.isSlidingHorizontally() > 30)
+    {
+        onScrollRightFunc();
+        touch.resetScrollHorizontal();
+    }
+
+    if(onScrollLeftFunc && touch.isSlidingHorizontally() < -30)
+    {
+        onScrollLeftFunc();
+        touch.resetScrollHorizontal();
+    }*/
+
+    for (int i = 0; i < this->children.size(); i++)
+    {
+        children[i]->update();
+    }
+}
