+++ lib/lua/src/lua_window.cpp
@@ -1,0 +1,7 @@
+#include "lua_window.hpp"
+
+LuaWindow::LuaWindow()  
+{
+    widget = new gui::elements::Window();
+    init(widget);
+}
